+++ objs/objs.cc
@@ -1048,7 +1048,7 @@ bool Game_object::edit(
 	        cheat.in_map_editor()) {
 		editing = 0;
 		Tile_coord t = get_tile();
-		unsigned long addr = reinterpret_cast<unsigned long>(this);
+		uintptr addr = reinterpret_cast<uintptr>(this);
 		std::string name = get_name();
 		if (Object_out(client_socket, Exult_server::obj,
 		               addr, t.tx, t.ty, t.tz,
@@ -1073,7 +1073,7 @@ void Game_object::update_from_studio(
     int datalen
 ) {
 #ifdef USE_EXULTSTUDIO
-	unsigned long addr;
+	uintptr addr;
 	int tx, ty, tz;
 	int shape, frame, quality;
 	std::string name;
